requirements.txt
numpy==1.18.0
opencv-python==4.1.2.30
Pillow==6.2.1
PyAutoGUI==0.9.48
pywinauto